pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "market-data-pipeline"
version = "0.1.0"
description = "Market data ingestion and validation pipeline"
authors = [{ name = "Andrew Geday" }]
dependencies = [
  "pandas",
  "numpy",
  "yfinance",
  "pyyaml"
]